Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
av==13.1.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloup==3.0.6
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
fastapi==0.115.11
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2025.2.0
future==1.0.0
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
isosurfaces==0.1.2
Jinja2==3.1.5
jiter==0.8.2
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
openai==1.65.2
packaging==24.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyglet==2.1.3
pyglm==2.8.0
Pygments==2.19.1
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
screeninfo==0.8.1
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.6
srt==3.5.3
starlette==0.46.0
svgelements==1.9.6
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==14.2
whisper.ai==1.0.0.1
annotated-types==0.7.0
anyio==4.8.0
av==13.1.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloup==3.0.6
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.11
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2025.2.0
future==1.0.0
glcontext==3.0.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.81.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==1.3.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
groq==0.19.0
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
isosurfaces==0.1.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.45
langchain-text-splitters==0.3.6
langsmith==0.3.15
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
openai==1.65.2
orjson==3.10.15
packaging==24.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyglet==2.1.3
pyglm==2.8.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
screeninfo==0.8.1
shapely==2.0.7
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
srt==3.5.3
starlette==0.46.0
svgelements==1.9.6
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==14.2
whisper.ai==1.0.0.1
zstandard==0.23.0